backend/requirements.txt
django==5.1.9
requests
google-genai
Pillow
